+++ include/GafferUI/GraphGadget.h
@@ -38,9 +38,9 @@
 #pragma once
 
 #include "GafferUI/ContainerGadget.h"
+#include "GafferUI/ContextTracker.h"
 
 #include "Gaffer/CompoundNumericPlug.h"
-#include "Gaffer/Context.h"
 #include "Gaffer/Plug.h"
 
 #include "Gaffer/BackgroundTask.h"
@@ -54,11 +54,6 @@ namespace Gaffer
 IE_CORE_FORWARDDECLARE( Node );
 IE_CORE_FORWARDDECLARE( ScriptNode );
 IE_CORE_FORWARDDECLARE( Set );
-}
-
-namespace GafferUIModule
-{
-	class ActivePlugsWrapperClassToUseAsFriend;
 }
 
 namespace GafferUI
@@ -214,10 +209,6 @@ class GAFFERUI_API GraphGadget : public ContainerGadget
 		void rootChildRemoved( Gaffer::GraphComponent *root, Gaffer::GraphComponent *child );
 		void selectionMemberAdded( Gaffer::Set *set, IECore::RunTimeTyped *member );
 		void selectionMemberRemoved( Gaffer::Set *set, IECore::RunTimeTyped *member );
-		void updateFocusPlugDirtiedConnection();
-		void focusChanged();
-		void focusPlugDirtied( Gaffer::Plug *plug );
-		void scriptContextChanged( const Gaffer::Context *context, const IECore::InternedString& );
 		void filterMemberAdded( Gaffer::Set *set, IECore::RunTimeTyped *member );
 		void filterMemberRemoved( Gaffer::Set *set, IECore::RunTimeTyped *member );
 		void inputChanged( Gaffer::Plug *dstPlug );
@@ -268,14 +259,10 @@ class GAFFERUI_API GraphGadget : public ContainerGadget
 		Gaffer::NodePtr m_root;
 		Gaffer::ScriptNodePtr m_scriptNode;
 		RootChangedSignal m_rootChangedSignal;
-		IECore::MurmurHash m_scriptContextHash;
 		Gaffer::Signals::ScopedConnection m_rootChildAddedConnection;
 		Gaffer::Signals::ScopedConnection m_rootChildRemovedConnection;
 		Gaffer::Signals::ScopedConnection m_selectionMemberAddedConnection;
 		Gaffer::Signals::ScopedConnection m_selectionMemberRemovedConnection;
-		Gaffer::Signals::ScopedConnection m_focusChangedConnection;
-		Gaffer::Signals::ScopedConnection m_focusPlugDirtiedConnection;
-		Gaffer::Signals::ScopedConnection m_scriptContextChangedConnection;
 
 		Gaffer::SetPtr m_filter;
 		Gaffer::Signals::ScopedConnection m_filterMemberAddedConnection;
@@ -315,35 +302,11 @@ class GAFFERUI_API GraphGadget : public ContainerGadget
 
 		GraphLayoutPtr m_layout;
 
-		// Track if we need to run an active state update.  If true, then if there isn't already an
-		// m_activeTask running, we need to start one.  ( Helpful to track separately so we know
-		// we need to restart if the task gets cancelled somehow )
-		bool m_activeStateDirty;
-		void dirtyActive();
+		void applyFocusContexts();
 
-		// Used to run updateActive()
-		std::shared_ptr<Gaffer::BackgroundTask> m_activeStateTask;
+		ContextTrackerPtr m_contextTracker;
+		Gaffer::Signals::ScopedConnection m_contextTrackerChangedConnection;
 
-		// Does the actual calculation of the active state, then calls applyActive.
-		// Should be run on background thread
-		void updateActive();
-
-		// Applies the active state to the child Gadgets ( must be called on UI thread )
-		void applyActive(
-			std::shared_ptr< std::unordered_set<const Gaffer::Plug*> > activePlugs,
-			std::shared_ptr< std::unordered_set<const Gaffer::Node*> > activeNodes
-		);
-
-		// Given a plug and context, returns all nodes and plugs which contribute to evaluating that
-		// plug
-		static void activePlugsAndNodes(
-			const Gaffer::Plug *plug,
-			const Gaffer::Context *context,
-			std::unordered_set<const Gaffer::Plug*> &activePlugs,
-			std::unordered_set<const Gaffer::Node*> &activeNodes
-		);
-
-		friend GafferUIModule::ActivePlugsWrapperClassToUseAsFriend;
 };
 
 IE_CORE_DECLAREPTR( GraphGadget );