+++ Changes.md
@@ -13,6 +13,7 @@ Improvements
   - Added `ignoreMissingAlpha` plug.
   - Optimised the pass-through of the alpha channel.
 - EditScope : Filtered the EditScope menu to show only nodes that are active in the relevant context.
+- GraphGadget : Improved highlighting of active nodes, with more accurate tracking of Loop node iterations.
 
 Fixes
 -----